requirements.txt

openai>=1.12.0


chromadb>=0.4.22
sentence-transformers>=2.3.1
sqlite3  # Built-in


click>=8.1.7
rich>=13.7.0
prompt-toolkit>=3.0.43


psutil>=5.9.8
pyperclip>=1.8.2


selenium>=4.16.0
playwright>=1.41.0


python-dotenv>=1.0.0
pyyaml>=6.0.1
pydantic>=2.5.3
tenacity>=8.2.3


structlog>=24.1.0


pytest>=7.4.4
pytest-asyncio>=0.23.3
